Full wiki setup
These files introduce the 'wiki' role, which can do the wiki setup with
salt automatically. Only 3 things need to be done manually:
- create /srv/www/*.opensuse.org/secrets.php (with $wgSecretKey and
  $wgDBpassword)
- import (and upgrade) the database
- import the uploaded files

secrets.php can be automated in the future (when we have handling
passwords in salt in place), but I don't plan to wait for this ;-)

The IPs in the wiki pillar for elasticsearch_server and mysql_server are
those from my local testing and obviously need to be adjusted for the
openSUSE setup.

Speaking about local testing - I completely tested this locally with a
set of VMs and therefore expect that it will "just work[tm]" ;-)
(as soon as all my pending merge requests are accepted)