Don't set $wgCookieDomain to fix login/edit problems
Trying to edit el.o.o or pl.o.o, I always got an error message:
    Sorry! We could not process your edit due to a loss of session data.
    You might have been logged out. Please verify that you're still
    logged in and try again. [...]"

The reason for this was $wgCookieDomain = 'opensuse.org'.

el.opensuse.org includes MediaWiki:Common.css from en.opensuse.org and
gets the english session cookies this way. Since the CSS is loaded after
the HTML and the latest cookie wins, the next request sends the english
cookies to the el wiki

The solution is obviously to restrict $wgCookieDomain to en.opensuse.org
etc. Luckily this is the default behaviour, so we can simply delete the
line that sets $wgCookieDomain.