Add support to ignore non-existing states in {salt,pillar}/top.sls
In the past we needed to create a state that it was included in top.sls,
even if it was empty. By adding the include_optional macro, we can now
check if the file exists on the disk and then include it. This way we
can get rid of all the empty files.