bin/get_formulas.py: don't verify the SSL cert for the gitlab.i.o.o repos
We are using a freeipa created cert for gitlab.i.o.o, thus the
containers created by CI runner can't verify it. This creates the issue
that we can't fetch the remote. Since we plan to switch to letsencrypt
wildcards as soon as they are available, it's easier for now to pass an
environment variable to disable the SSL verification for this command