add monitoring:check_zypper:whitelist to pillar/id/*
The whitelists are based on today's check_zypper output (without using
the ignore list) and should therefore not contain any outdated entries.

Note that the ignore list is wrapped in an {% if osrelease == ... %}
condition. This will hide the whitelist after a distribution upgrade
to ensure we don't keep outdated entries forever.
If we still need to whitelist packages after the upgrade, we'll have to
update the version check (after cleaning up the whitelist).