fix osem sudo config and allow osem-admins group
The sudo rules for the osem user need to be quoted in salt to avoid the
"... PASSWD:" part is interpreted as key, and everything after it as
value (which will result in an invalid sudo rule).

Also add group permissions for the osem-admins group so that they can
use sudo.