Drop server aliases for cs.o.o and zh-tw.o.o
Since $unknown_time,
- cz.opensuse.org results in "Unconfigured host name"
- zh_tw.opensuse.org doesn't exist in DNS

Nobody complained, and that means both aliases are probably superfluous ;-)

These two were the only wikis with aliases, which means we no longer
need ServerAlias in the wiki apache config. Therefore this commit also
drops support for that. If we ever need it again, we can easily re-add
it - or simply setup a redirect in haproxy instead.