pagure: Set default branch for new projects to be "main"
As part of being an inclusive community, we want to try to avoid
problematic terminology that may implicitly make people feel unwelcome.

The Git upstream project is in the process of migrating the default
branch name from "master" to "main", and there's no reason for us to
not be ahead of the curve here, so let's set this.