Test only nginx in nginx test
The role test is poorly designed and causes several errors. If testing
of the complete web roles is desired, a test in the form of "test_web"
or similar should be introduced. This test, "test_nginx" is hence
changed to only test the nginx configuration, matching what other
software specific tests, such as "test_haproxy" are doing.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>