Add OpenVPN login log script
This adds a script which is called for every OpenVPN client connection,
writing the IP address and timestamp into a file.
Needed to, after this has been installed for some time, find and revoke
users who no longer use the service.
Additionally it allows for assessing whether an account reported as
compromised has been used.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>