Blackhole os-internal in legacy locations
Until we fully roll out IPv6 in all locations, we need to ensure
internal connectivity for core services, such as LDAP.

Machines in legacy sites generally communicate fine with internal
services in PRG2 using IPv4, but only if the machine in question does
not have other routes for IPv6, as is the case with machines having a
a public IPv6 address and matching default gateway assigned.
Such machines will try to route traffic targeting hosts inside os-internal
through the internet, as the GUA prefix does not differentiate between
private and public addresses. As we, by design, do not allow internal traffic
to be routed through the internet, we need to ensure affected machines
fall back to IPv4, which they have an internal route for.

This patch adds a blackhole route to os-internal on affected machines,
forcing all related traffic to hosts in the os-internal network to be
sent using IPv4. This is especially needed for applications not
supporting a "-4" flag or similar address family configuration options.

Alternatives considered:
 - /etc/hosts entries for affected domains
   -> difficult to keep track of, leads to "surprises" down the line
 - additional DNS names for affected domains with, for example, a
   "legacy-" or "ipv4." prefix
   -> breaks TLS validation if this additional name is not contained in
   the certificate SANs

Both alternatives additionally require all affected services/domains to
be equipped with the respective additional entries.

The chosen solution, blackholing, was henced deemed the most effective, and
the most easy to maintain - once full IPv6, including internal routing to all
sites, is rolled out, we can easily revert this patch again.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>