Refactor HAProxy http-request/http-response pillar
Use native template logic instead of "extra". This is not only easier
to read, but additionally places the blocks in the right location in
the generated configuration file, resolving the warnings

```
a 'http-request' rule placed after a 'redirect' rule will still be processed before.
```

which was printed for each affected line.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>