Remove Btrfs conditional
Execution modules called in the pillar are executed on the master,
this if-clause does not have the desired effect.
Since the Snapper feature does not seem to cause any errors if enabled
on non-snapper capable machines (i.e. ones with a non-Btrfs root filesystem),
it should be safe to omit the conditional and have the option set on
all minions.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>