Drop upstream formulas test
Test is useless, we are using downstream mirrors, sometimes forks, of
the formulas which are already tested as part of "show_highstate".
Additionally, the upstream formula test fails if we use downstream patches
containing files not available in the upstream repository, as state
includes fail to find them.
As a replacement for this, a new test was introduced in salt-formulas-git,
which correctly tests the downstream formulas when they get updated.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>