Remove tuned
This used to be included in the Kiwi image, but no longer is.
On the machines it is installed, it is either only idling,
without being configured with any profile, or configured with
"virtual-guest", which merely sets vm.dirty_ratio and vm.swappiness,
settings which, if at all, should be managed using the sysctl pillar.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>