Disable truthy rule for reboot_safe
We use "yes" and "no" values for the reboot_safe grain, which makes
sense for human readability, albeit the values internally being
converted to "true" and "false".
To avoid confusion with users inspecting the valid_custom_grains.yaml
file to check which values are valid, the rule is disabled for this
situation, instead of changing the values to "true" and "false".

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>