test_infra_data: allow complicated function
Generally we do not allow functions which branch out more than a
set amount of times, to improve readability and to help others
with understanding the logic.
In this case it is not feasible to refactor the script, as the
duplication logic considers different factors.
Hence, the respective rule is set to be ignored here.

To make it slightly better, some spacing is introduced.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>