Render pillar files related to failed state files
A common reason for state files failing to render are references
to pillars which do not provide the expected values.
Since we make heavy use of pillar templating, it is not always
easy to find where unexpected values originate.
This patch aims to make this easier, by extracting and rendering
pillars referenced in state files which failed to render.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>