Add certificate conditional to dehydrated.target
Only render states if certificates are assigned to the respective
host, otherwise the test suite will fail as the "cert" user is
currently only installed in the test setup script for roles making
use of role.common.certificate.
If we use this for more roles in the future, we can move the cert
user setup snippet to a common place and subsequently remove this
conditional.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>