Configure Salt LDAP authentication
Used for ACL based privilege delegation and API authentication.
Initially, a simple ACL for our "wheel" group is included, allowing
users in it to use the Salt same way as root.
Initially this will only work for "salt -aldap ..." and with the API.
In a later patch, regular CLI use will be facilitated, and further
fine grained rules will be added.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>