Move certificates pillar to common
Avoid the need to include this on a per role basis, allow the
infra/certificates/*.yaml to be operated without requiring error
prone manual consideration of pillar includes.
Since the conditionals in the pillar template check for whether
certificates are needed on any given minion, this will not introduce
any changes to minions not listed as certificate deployment targets.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>