Enable debug Master/Minion logging in test suite
It is very useful being able to inspect the rendered pillar and state
templates as well as debug logging messages implemented in our Jinja
code.
To reduce the amount of useless log messages printed alongside,
filtering is applied before the artifact file is generated.
Additionally, the salt/salt-call command line output is moved to
a cli.txt file to reduce noise in the main CI output.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>