Strip comments from Dehydrated known hosts
The Salt known_hosts parsing logic does not support comments.
It finds the algorithm column by removing the last two fields of
a given known_hosts line, which breaks if the line contains a
comment column, causing the state to not be idempotent due to
existing entries not being detected. Solve this by only feeding
the second column, which should always contain only the key part,
to the known hosts state.
The comment is not needed anyways, as the FQDN is already contained
in the host column.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>