Send Paste logs to stdout
This allows its logs to be collected in the systemd journal instead
of in a file, reducing the need for logrotate configuration and
facilitating automatic forwarding to our syslog server.

Additionally, make the state idempotent by removing module calls
and having the service states watch for configuration file changes
by themselves.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>