Increase maximum libvirt client requests
Repeatedly, the following message was printed by either of
libvirtd, virtlockd, or virtlogd:

```
Client hit max requests limit 1.
This may result in keep-alive timeouts.
Consider tuning the max_client_requests server parameter
```

The number mentioned in the message seems to not be meaningful,
it does not match our existing setting.
The limit is now increased even further, especially to facilitate
busy live migration scenarios.
Additionally, the maximum workers are increased - the documentation
suggests to set the maximum requests to be a fraction of the maximum
workers, but also suggests to have the maximum workers equal the
maximum clients.
Since the formula defaults define the maximum clients to 128, we
align the workers with this count and double the maximum requests
to make them a fraction.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>