Correct UPDATE_CMD conditional for Leap
The "osfullname" grain resolves to "Leap", not "openSUSE Leap".
Since we apparently run the default UPDATE_CMD (i.e. "zypper up")
with no issues on all our Leap machines, we correct this but add
a second conditional to only apply it for machines which are not
safe to reboot. This will help with reducing unexpected maintenance
of cluster machines which may not have immediately needed a
non-security relevant update, we can assess on non-cluster/reboot
safe machines first.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>