Manage /etc/sysconfig/services
This manages the file and disables service restarts on updates
for minions with the "hypervisor.cluster" or "mariadb" role.

The "hypervisor.cluster" role is because updates of packages contained
in the SLE HA stack are officially not safe to be installed without
stopping the cluster stack prior to an update:
https://documentation.suse.com/sle-ha/15-SP5/html/SLE-HA-all/cha-ha-migration.html#sec-ha-migration-update

The "mariadb" is because we yet need to implement health checks to
assess whether one Galera node finished synchronizing before the next
one is restarted.

Unfortunately our RPM service macros only check this one variable
and do not allow for individual services to be excluded, which would
be preferable.

Signed-off-by: Georg Pfuetzenreuter <georg.pfuetzenreuter@suse.com>