Allow skipping of reloads in certificate hook
While bootstrapping a new machine which had a lot of existing
certificates assigned to it, I noticed the first deployment to
trigger a lot of reloads.
This is not only inefficient in the situation, but can also cause
errors if the service fails to reload due to other certificates
not yet being present.
Hence introduce a SKIP_RELOAD variable, which can be transiently
set by the administrator to skip the reload/restart logic.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>