Set Redis groups as system groups
The Redis package installs a system group named "redis", however the
users formula states, included via role.base, which we use to add users
as members to the "redis" group, are executed before the Redis
package installation, hence they create the matching group before the
package is able to.
To avoid the users formula creating a non-system "redis" group
on systems on which the Redis package is not yet installed, set the
"redis" group to be a system group explicitly.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>