Refactor HAProxy stats sockets
Improve security by splitting sockets into a read-write ("operator")
and a read-only ("user") one.
The Keepalived check script is adjusted to make use of the stats-ro
socket, which it has access to using the "haproxy" group.
Administrators can use the stats-ro socket for querying statistics,
and the stats-rw socket to perform non-destructive write operations,
such as counter resets.
No full "admin" socket is implemented at this time, as we do not have
a use case for live configuration changes.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>