Allow repository access in rootless CI containers
GitLab CI clones the repository as root, but we prefer to use a
regular user for some of our job specific containers.
Add a helper script and call it before the respective script
executions to work around the "dubious ownership" error as there
does not seem to be a GitLab CI solution to clone as the target
user.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>