Saltify Kanidm Client
This is the client configuration to convert clients from SSSD to Kanidm. This will
be rolled out in phases. This initial phase is simply adding the packages and
configuration to hosts to detect any failures to install or communicate to the
server.

Phase 2 will be changing pam/nsswitch/sshd_authorised keys to Kanidm, while leaving
SSSD in place as a backout option.

Once we confirm that everything works as expected, Phase 3 will be cleanup and
removal of SSSD from all hosts.