fetch-image: signature download etag hack
The ETag functionality in Salt theoretically works to assess whether
a file needs to be re-downloaded. For its operation, it caches an
".etag" file along with the desired file in the minion cache.
If the ".etag" file is missing, for example if the minion cache got
emptied, it should always trigger a re-download.

It was found that this seems to not always correctly work.
In our fetch-image states, the checksum file got correctly re-downloaded
upon its change on the server, but the signature did not - particularly,
if the minion cache was empty and no ".etag" was present.
This causes the new checksum to be verfied against an old signature,
which naturally fails.
It is not clear why the logic works for one of the two file but not
for the other.
To work around this issue, we assess whether the cache is populated
at render time. Should the cache be empty, we omit the ETag logic
for the problematic file, and always re-download it.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>