Increase mailman3 resources
Increase memory to 32GB - mailman-web suffers from a fundamental
issue with memory usage:
https://gitlab.com/mailman/mailman/-/issues/1050
Increasing the memory of the machine is not a good solution, but
is feasible as an attempt to decrease the amount of out of memory
kills.
Once an upstream solution gets implemented, the machine memory
should be decreased again.

Increase CPU cores to 16. The mailman-web gunicorn daemon is
configured with 16 workers (the default). The documentation
recommends workers=(2*cores)+1:
https://docs.gunicorn.org/en/latest/design.html#how-many-workers
This would make for a core count of 9. For good measure, we
generously round this up to 16, to not only satisfy gunicorn better,
but also to give room to the various other processes running on the
machine.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>