Split out service pattern in schemas
As a result of a previous patch, macros on certificates can now be
combined with additional services. To facilitate this in the schema
test, it would have become needed to copy the service pattern to
the "macro" section. Since this would have been a third copy of the
same pattern, it is instead split out to a separate "service" schema,
and then referenced in the needed places. This avoids unnecessary
code duplication and makes it easier to amend the pattern in the
future by not having to track down multiple occurrences.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>