Drop get_formulas.py
No longer useful since Salt no longer allows states to be installed
outside the file root - and the /srv/formula directory populated by
the script is not mountable as a file root itself given the states
residing in subdirectories under the respective formula repository
directories.
The installation of Git based formulas in the test environment is
taken care of by clone_formulas.sh now, which populates /srv/formulas
with a checkout of salt-formulas-git.git, which in return was equipped
with top level state directory symlinks.

https://gitlab.infra.opensuse.org/infra/salt/-/issues/20
https://bugzilla.opensuse.org/show_bug.cgi?id=1219985

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>