Drop master formula logic
This refactors the Salt Master role to only dumbly clone the
salt-formulas-git repository instead of dynamically linking the
repository contents. Linking to directories outside the ones listed
as file_roots is no longer possible, and we need to maintain all
links inside the repository, for which a separate patch there will
be submitted.

https://gitlab.infra.opensuse.org/infra/salt/-/issues/20
https://bugzilla.opensuse.org/show_bug.cgi?id=1219985

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>