Inhibit Prometheus email alerts at night
We install automatic OS updates which come with service restarts
and reboots at night. Avoid irrelevant mails for breakfast but
still have the alerts available in IRC.
This change comes with a refactoring of the single receiver to
a routes based approach which may also come useful for other receiver
specific settings in the future.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>