Define zypp connection timeout
This is an attempt to make zypp(er) fail early if no connection to
a download server can be established, as otherwise any Salt operation
involving packages freezes for a long time.
In our internal infrastructure it should be reasonable to expect a
fast server response and to assume a deeper problem if not the case.
Hence this lowers the timeout from the default 60 to 10 seconds.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>