Rework wanted Prometheus components
In common.monitoring a "wanted" section for Node Exporter was
introduced, which uses a dictionary based approach to override the
package name.
Prometheus/Alertmanager do not require such override and hence used
the list based approach in the past, but are now changed to use a
dictionary in order to merge with Node Exporter, ensuring a monitoring
master gets both, the server and the exporter components, installed.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>