Add empty Prometheus receiver
dfc4854d0ee992d664d92ea50f258f92b7477a20 introduced more alerting
routes, including one for informational alerts which only end up
in the dashboard. This patch adds the respective receiver, an "empty"
one without any media configuration to act as a "dashboard-only"
target (it is technically not linked to the dashboard, which receives
all alerts anyways - it is more a "stopper" for alerts which should
not continue its travels to other communication channels).

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>