Disable Wicked firewall hook
If a host uses firewalld, the zone-interface mappings are
explicitly defined in the respective firewall zone using the
firewalld pillar. Wicked was observed overwriting the zone defined
mapping with the default "public" zone on boot, naturally breaking
connectivity. Disable this behavior by disabling the firewall
hook for all VM interfaces.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>