Refactor grains in mine
For differentiating between physical and virtual machines in the
pillar, currently needed for grouping monitoring targets, the
"virtual" grain needs to be available in the mine.
To make fetching of multiple grains in the mine easier, as well
as to avoid having to define more mine functions if other grains
are of interest in the future, the existing "fqdn" mine is refactored
to a generic "grains" mine, containing the "fqdn" and "virtual" grains
in a single dictionary per minion.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>