Refactor Prometheus scrape targets
This employs the mine to gather minions and to pre-sort them into
monitoring "groups", allowing us to automatically enroll minions
into both node and service specific monitoring monitoring without
the need to explicitly list them in the monitoring pillar.
It additionally allows for easy labelling of scrape targets using
data available in the mine, for example to split physical and virtual
machines or to label clusters.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>