Enable HAProxy Prometheus metrics
This adds a listener on port 8404 to all HAProxy nodes, exposing
an internal /metrics endpoint which can later be scraped by
Prometheus.
The metrics backend on Atlas needed to be renamed, as it is not
possible to have a backend with the same name as a listener, and
the decision to name the metrics listener "metrics" was already
made before noticing.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>