Simplify Prometheus exporter group/role mapping
Avoid having to declare a conditional for every new role equipped
with Prometheus exporters. Implement a dictionary containing
target groups and the respective Salt roles. This allows for easier
definition of new groups and easier mapping of a group to multiple
Salt roles whilst also facilitating minions with multiple roles
mapping to the same exporters.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>