Prepare node exporters for textfile collection
In order to facilitate collection of metrics for applications which
do not expose any natively, as well as to replace legacy nsca based
monitoring hooks, minions are prepared for textfile collection using
the already installed Prometheus Node Exporter.
For this, we deploy a canonical directory which textfile metrics can
be placed in, and install moreutils to have `sponge` available for
writing them.
Given the low overhead this is deployed to all minions.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>