Push os-public through OpenVPN
In 7876bac9b553c4d72b93be7ec61ebdec19788984 the firewall rules
were extended to grant VPN clients the same access to hosts in
os-public as over the internet.
Building on this, we push the os-public network as a route to
VPN clients, making public services routed through the VPN if
connected. This allows for access to network restricted areas
of certain public services.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>