Configure Promact and MySQL processlist action
To gain better visibility into the cause of alerts, Promact
is deployed to act on webhooks received by Alertmanager.

As a first action, this is configured to save the MySQL process list
to a directory served under http://monitor.i.o.o upon alerts about
too many MySQL connections, allowing for inspection of the database
status as it was at the time of the event.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>