Create /usr/local/libexec on Leap
This directory is already shipped by default on Tumbleweed where
it is used as the local equivalent of %_libexecdir -
this patch implements a "FHS backport" to avoid having to create
the parent directory in states using it on Leap.

Background for starting to use /usr/local/libexec to begin with
was the desire for a more conforming place than /usr/{,local/}lib
(the current standard on Leap for %_libexecdir, albeit not quite
matching the FHS description) to store scripts intended for the
execution through other, parent, applications
(i.e. ones not intended to be available in PATH).

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>