Simplify etckeeper profile
On initial highstates, the etckeeper init state would fail due to
git-core not yet being installed.
Since git-core is listed in the common pillar and managed through the
zypper formula, add a dependency to its installation state in the
respective etckeeper ones.
To streamline the setup, move the etckeeper packages to the common
pillar as well, and to simplify it further, remove the obsolete
SLE 12 conditional.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>