Set kernel hostname on suse_ha nodes
The patch in afc8249275e6fd95552463da7aa60c72eb04542b was rejected
in favor of an /etc/hosts approach. The latter seems to work reliably
on VMs, but does not on physical machines. This is an issue with
SUSE HA, since the transient host name does not match the expected
cluster node name (the FQDN), causing "member weirdness errors" and
election loops.
Hence force the kernel hostname to match the Salt ID, which in our
case is always ensured to match the FQDN, which in return matches
our configured cluster node names.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>