Switch tcpka in HAProxy backends to srvtcpka
As per the HAProxy documentation:

" ... it is strongly recommended to explicitly use "option clitcpka" and
"option srvtcpka" when the configuration is split between frontends and
backends. "

https://www.haproxy.com/documentation/haproxy-configuration-manual/2-9r1/#4-option%20tcpka

Rename existing occurrences and remove ones which do not make sense
(backends exclusively serving a local error page without
connecting to any backend servers).

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>