Improve zypper metrics resiliency
The metrics script could return invalid metrics, such as
zypper_pending_updates without a value, causing errors in the
textfile collection if a zypper execution failed (for example.
due to not being able to aquire a lock).

Address this by writing each metric only if the corresponding zypper
operation succeeded, and preventing alteration of the metrics file
if a failure occured.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>