Configure smartd on physical machines
Manage smartd on all physical machines detect disk failures early.
The schedule default is applied from the SUSE packaged defaults,
notifications are sent to the local root mailbox and forwarded
to admin-auto@ through aliases.
To avoid superfluous scanning of devices (especially on clusters
with many multipath disks) we declare the relevant local disks on
an id or cluster basis.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>