Configure SMART alerts
Notify us about sad disks.

It would be good to have alerting for some individual attributes
(for example SSD wear level) as well, but there does not seem to
be consistency between the models we use, hence only alert about
temperature and the general SMART status here and leave the rest
to smartctl's built in email notifications.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>