Add additional port to Pagure SSH setup
Merely the ListenAddress setting will cause the listen addresses
and ports from the included main sshd_config to be considered as
well, causing the sshd-public daemon to bind and occupy port 22
if started earlier (if started later, it would error but continue,
hence adding systemd ordering would be an alternative).
Using Port 2222 and only specifying the hostname with ListenAddress
does not suffice either, as it would try to bind the primary
interface address in addition to localhost.
A clean solution would be to write a config with common options
and then to include that in both the main sshd and the public sshd
configurations, which by themselves would only specify the differing
options like ListenAddresses. But since it would require refactoring
since the Salt formula for SSH currently only supports writing
a regular sshd_config file, rather work around it by adding an
additional port through "Port", which opens another superfluous
listener but disengages the one from the main configuration file.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>