Fixate X-Forwarded-Proto for Pagure
The $scheme variable evaluates to "http", since our backend
connection is only using plain text HTTP. This would overwrite
the X-Forwarded-Proto header set to "https" by HAProxy on Atlas
with "http", causing issues in the backend application due to it
generating some URLs with a different schema than the one used
by the connecting client.

https://progress.opensuse.org/issues/161273

Simply removing the option (in an attempt to have it preserve the
header and value set by HAProxy on Atlas, which would be cleaner
than overwriting it with a hardcoded value) did not suffice, but
the presented solution should not cause issues due to us enforcing
HTTPS client connectivity through a permanent redirect anyways.

Consider removing this when implementing backend TLS connectivity.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>