Lower mine_interval
The mine update is very expensive due to us collecting
state.show_states, which is used in the monitoring pillar which
can as a result be incomplete during the long running updates.
Since the data we collect in the mine rarely changes anyways,
and if it does, is usually accompanied by administrative action,
set the interval to 12 hours instead.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>