Set root home directory mode
Some minions include authorized keys for the root user through
pillar/ssh_keys/ utilizing the users formula. The users formula
sets the home directory mode for users declared in its pillar
to 750. This conflicts with our chkstat/permctl security policy,
which enforces a mode of 700 on /root. Hence explicitly set the
home directory mode for the root user to 700 in the users pillar
as well.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>