Configure nginx-next repository on download.i.o.o
We use the fancyindex module for more comfortable browsing of
the internal mirror but it is not part of the distribution.
It needs to be installed with the same nginx version it was built
against, but we don't want to ship a newer nginx to all machines.
Hence add a repository linking the packages from server:http for
this purpose.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>