Except WireGuard port from UDP rejection
In zones/1204_os-public.nft we already allow inbound WireGuard
traffic from provo-gate to asgard* to allow for connection
initiation in both directions.
To make this work, the port additionally needs to be excluded in
the UDP rejection rule which, being in the global zone, has
precedence over segment specific rules.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>