Add default 'mydestination' to fix delivery of cron etc. mails
... to make sure that all VMs feel responsible for mails cron sends to
a user @ their own hostname, and applies /etc/aliases (typically to
forward the mails to admin-auto)

The chosen value is also the postfix default value for mydestination.

Partially fixes poo#156502 (some machines also need master.cf fixes, but
that's another can of worms.

While on it, sort the maincf pillar.