Do not log Apache exporter status requests
The Apache Prometheus exporter frequently requests the status
information from httpd on localhost. This is expected and not
interesting to know about in the access log - except these requests
from being logged and reduce disks filling up with it.

The pattern matches the following sample request:

```
::1 - - [...] "GET /server-status/?auto HTTP/1.1" 200 985 "-" "Prometheus-Apache-Exporter/1.0.7"
```

d+ is used to facilitate potential double digit version numbers
in the future.

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>