Do not log HAProxy health check requests in Apache
The HAProxy reverse proxies frequently query the backend httpd
servers to assess their health. This is expected and not relevant
in the backend access logs - except these requests from being
logged and reduce disks filling up with it.

The pattern matches the following sample requests, and considers
reverse proxies in both the os-internal and the os-public networks:

```
2a07:de40:b27e:1204::11 - - [...] "OPTIONS / HTTP/1.1" 200 - "-" "haproxy/health-check"
2a07:de40:b27e:1203::11 - - [...] "GET / HTTP/1.1" 200 - "-" "haproxy/health-check"
```

Signed-off-by: Georg Pfuetzenreuter <mail@georg-pfuetzenreuter.net>